refactor: 안쓰는 폴더 삭제, DialPad 폴더 위치 이동, fragment 삭제
