+++ frontend/src/hooks/useDialPad.ts
@@ -1,6 +1,6 @@
 import { PHONE_NUMBER_LENGTH, REGEX, ROUTER_PATH } from '../constants';
 import { ChangeEvent, KeyboardEvent, useRef, useState } from 'react';
-import { DialKeyType } from '../components/Dialpad';
+import { DialKeyType } from '../pages/EnterPhoneNumber/Dialpad';
 import { useLocation, useNavigate } from 'react-router-dom';
 
 const addHypen = (phoneNumber: string) => {
