+++ frontend/src/pages/Admin/RewardPage/style.tsx
